Some awesome separation of concerns
