+++ packages/cst-to-ast/src/constants.ts
@@ -1,0 +1,19 @@
+
+export enum MathMode {
+  ALWAYS,
+  PARENS_DIVISION,
+  PARENS,
+  STRICT_LEGACY
+}
+
+export enum RewriteUrlMode {
+  OFF,
+  LOCAL,
+  ALL
+}
+
+export enum StrictUnitMode {
+  LOOSE,
+  WARN,
+  ERROR
+}
