+++ packages/cst-to-ast/src/tree/nodes/declaration.ts
