+++ packages/cst-to-ast/src/tree/nodes/extend.ts
