+++ packages/cst-to-ast/src/tree/nodes/negative.ts
